@@ -1,6 +1,6 @@
 {
 	"name": "adaptivecards-designer",
-	"version": "0.5.0",
+	"version": "0.7.1",
 	"description": "Adaptive Card designer embeddable control",
 	"author": "AdaptiveCards",
 	"license": "MIT",
@@ -13,26 +13,34 @@
 	],
 	"scripts": {
 		"clean": "rimraf build lib dist",
-		"prebuild": "tsc",
+		"copy-hostconfigs": "cpy ../../../samples/HostConfig/*.json ./src/hostConfigs",
+		"prebuild": "npm run copy-hostconfigs && tsc",
 		"build": "webpack",
 		"watch": "webpack --watch",
 		"start": "webpack-dev-server --open",
+		"start:preview": "webpack-dev-server --open-page previewFeatures.html",
+		"start:nohosts": "webpack-dev-server --open-page noHosts.html",
 		"dts": "dts-generator --name adaptivecards-designer --project . --out dist/adaptivecards-designer.d.ts",
 		"lint": "tslint -c tslint.json 'src/**/*.{ts,tsx}'",
 		"release": "npm run build && webpack --mode=production && npm run dts"
 	},
 	"dependencies": {
-		"adaptivecards": "^1.2.0-beta1",
-		"adaptivecards-controls": "^0.1.7",
-		"clipboard": "^2.0.1",
+		"adaptivecards": "^1.3.0-alpha.2",
+		"adaptivecards-controls": "^0.3.1",
+		"adaptivecards-templating": "^0.1.0-alpha.1",
+		"clipboard": "^2.0.1"
+	},
+	"peerDependencies": {
 		"monaco-editor": "^0.15.6"
 	},
 	"devDependencies": {
+		"cpy-cli": "^2.0.0",
+		"monaco-editor": "^0.17.1",
 		"rimraf": "^2.6.2",
-		"typescript": "^3.2.2",
-		"webpack": "^4.27.1",
-		"webpack-cli": "^3.1.2",
+		"typescript": "^3.5.3",
+		"webpack": "^4.38.0",
+		"webpack-cli": "^3.3.6",
 		"webpack-concat-plugin": "^3.0.0",
-		"webpack-dev-server": "^3.1.14"
+		"webpack-dev-server": "^3.7.2"
 	}
 }
