@@ -1,17 +1,5 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
-export * from "./containers/host-container";
-export * from "./card-designer-surface";
-export * from "./card-designer";
-export * from "./constants";
-export * from "./draggable-element";
-export * from "./designer-peers";
-export * from "./peer-command";
-export * from "./miscellaneous";
-export * from "./settings-manager";
-export * from "./side-pane";
-export * from "./toolbar";
-export * from "./tool-palette";
-export * from "./splitter";
-export * from "./treeitem";
+export * from "./adaptivecards-designer-standalone";
 export * from "./containers/index";
+
