@@ -5,11 +5,27 @@ import * as Controls from "adaptivecards-controls";
 import { DraggableElement } from "./draggable-element";
 import { IPoint } from "./miscellaneous";
 import * as DesignerPeers from "./designer-peers";
+import * as ACData from "adaptivecards-templating";
+import * as Shared from "./shared";
 
 export type CardElementType = { new(): Adaptive.CardElement };
 export type ActionType = { new(): Adaptive.Action };
-export type CardElementPeerType = { new(designerSurface: CardDesignerSurface, registration: DesignerPeers.DesignerPeerRegistrationBase, cardElement: Adaptive.CardElement): DesignerPeers.CardElementPeer };
-export type ActionPeerType = { new(designerSurface: CardDesignerSurface, registration: DesignerPeers.DesignerPeerRegistrationBase, action: Adaptive.Action): DesignerPeers.ActionPeer };
+export type CardElementPeerType = {
+    new(
+        parent: DesignerPeers.DesignerPeer,
+        designerSurface: CardDesignerSurface,
+        registration: DesignerPeers.DesignerPeerRegistrationBase,
+        cardElement: Adaptive.CardElement
+    ): DesignerPeers.CardElementPeer
+};
+export type ActionPeerType = {
+    new(
+        parent: DesignerPeers.DesignerPeer,
+        designerSurface: CardDesignerSurface,
+        registration: DesignerPeers.DesignerPeerRegistrationBase,
+        action: Adaptive.Action
+    ): DesignerPeers.ActionPeer
+};
 
 class DesignerPeerCategory {
     static Unknown = "Unknown";
@@ -98,22 +114,9 @@ export class CardElementPeerRegistry extends DesignerPeerRegistry<CardElementTyp
     }
 
     createPeerInstance(designerSurface: CardDesignerSurface, parent: DesignerPeers.DesignerPeer, cardElement: Adaptive.CardElement): DesignerPeers.CardElementPeer {
-        /*
-        var registrationInfo: IDesignerPeerRegistration<CardElementType, CardElementPeerType> = undefined;
-
-        for (var i = 0; i < this._items.length; i++) {
-            if (cardElement instanceof this._items[i].sourceType) {
-                registrationInfo = this._items[i];
-
-                break;
-            }
-        }
-        */
-
         var registrationInfo = this.findTypeRegistration((<any>cardElement).constructor);
 
-        var peer = registrationInfo ? new registrationInfo.peerType(designerSurface, registrationInfo, cardElement) : new DesignerPeers.CardElementPeer(designerSurface, this.defaultRegistration, cardElement);
-        peer.parent = parent;
+        var peer = registrationInfo ? new registrationInfo.peerType(parent, designerSurface, registrationInfo, cardElement) : new DesignerPeers.CardElementPeer(parent, designerSurface, this.defaultRegistration, cardElement);
 
         return peer;
     }
@@ -133,8 +136,7 @@ export class ActionPeerRegistry extends DesignerPeerRegistry<ActionType, ActionP
     createPeerInstance(designerSurface: CardDesignerSurface, parent: DesignerPeers.DesignerPeer, action: Adaptive.Action): DesignerPeers.ActionPeer {
         var registrationInfo = this.findTypeRegistration((<any>action).constructor);
 
-        var peer = registrationInfo ? new registrationInfo.peerType(designerSurface, registrationInfo, action) : new DesignerPeers.ActionPeer(designerSurface, this.defaultRegistration, action);
-        peer.parent = parent;
+        var peer = registrationInfo ? new registrationInfo.peerType(parent, designerSurface, registrationInfo, action) : new DesignerPeers.ActionPeer(parent, designerSurface, this.defaultRegistration, action);
 
         return peer;
     }
@@ -171,6 +173,8 @@ export class CardDesignerSurface {
     private _removeCommandElement: HTMLElement;
     private _peerCommandsHostElement: HTMLElement;
     private _lastParseErrors: Array<Adaptive.IValidationError> = [];
+    private _isPreviewMode: boolean = false;
+    private _sampleData: any;
 
     private updatePeerCommandsLayout() {
         if (this._selectedPeer) {
@@ -261,12 +265,45 @@ export class CardDesignerSurface {
                 this.onCardValidated(this._lastParseErrors, this.card.validateProperties());
             }
 
-            let renderedCard = this.card.render();
+            let cardToRender: Adaptive.AdaptiveCard;
+
+            if (this.isPreviewMode) {
+                if (Shared.GlobalSettings.previewFeaturesEnabled) {
+                    let cardPayload = this.card.toJSON();
+
+                    try {
+                        let template = new ACData.Template(cardPayload);
+            
+                        let context = new ACData.EvaluationContext();
+                        context.$root = this.sampleData;
+
+                        let expandedCardPayload = template.expand(context);
+
+                        cardToRender = new Adaptive.AdaptiveCard();
+                        cardToRender.hostConfig = this.card.hostConfig;
+                        cardToRender.parse(expandedCardPayload);
+                    }
+                    catch (e) {
+                        console.log("Template expansion error: " + e.message);
+                    }
+                }
+
+                if (!cardToRender) {
+                    cardToRender = this.card;
+                    cardToRender.designMode = false;
+                }
+            }
+            else {
+                cardToRender = this.card;
+                cardToRender.designMode = true;
+            }
+
+            let renderedCard = cardToRender.render();
 
             if (this.fixedHeightCard) {
                 renderedCard.style.height = "100%";
-            }
 
+            }
             this._cardHost.appendChild(renderedCard);
         }
     }
@@ -370,7 +407,7 @@ export class CardDesignerSurface {
             if (!peer) {
                 let registration = CardDesignerSurface.cardElementPeerRegistry.findTypeRegistration(Adaptive.AdaptiveCard);
 
-                peer = new registration.peerType(this, registration, action.card);
+                peer = new registration.peerType(peer, this, registration, action.card);
 
                 let parentPeer = this.findActionPeer(action);
 
@@ -436,7 +473,7 @@ export class CardDesignerSurface {
         this._designerSurface.style.width = "100%";
         this._designerSurface.style.height = "100%";
 
-        this._designerSurface.onkeydown = (e: KeyboardEvent) => {
+        this._designerSurface.onkeyup = (e: KeyboardEvent) => {
             if (this._selectedPeer) {
                 switch (e.keyCode) {
                     case Controls.KEY_ESCAPE:
@@ -685,6 +722,43 @@ export class CardDesignerSurface {
             }
 
             this.render();
+        }
+    }
+
+    get isPreviewMode(): boolean {
+        return this._isPreviewMode;
+    }
+
+    set isPreviewMode(value: boolean) {
+        if (this._isPreviewMode != value) {
+            this._isPreviewMode = value;
+
+            if (this._isPreviewMode) {
+                this._designerSurface.classList.add("acd-hidden");
+                this._dragHandle.renderedElement.classList.add("acd-hidden");
+                this._removeCommandElement.classList.add("acd-hidden");
+                this._peerCommandsHostElement.classList.add("acd-hidden");
+            }
+            else {
+                this._designerSurface.classList.remove("acd-hidden");
+                this._dragHandle.renderedElement.classList.remove("acd-hidden");
+                this._removeCommandElement.classList.remove("acd-hidden");
+                this._peerCommandsHostElement.classList.remove("acd-hidden");
+            }
+
+            this.renderCard();
+        }
+    }
+
+    get sampleData(): any {
+        return this._sampleData;
+    }
+
+    set sampleData(value: any) {
+        this._sampleData = value;
+
+        if (this.isPreviewMode) {
+            this.renderCard();
         }
     }
 }
