@@ -2,26 +2,30 @@
 // Licensed under the MIT License.
 import * as Clipboard from "clipboard";
 import * as Adaptive from "adaptivecards";
-import * as monaco from 'monaco-editor/esm/vs/editor/editor.api';
+import * as monaco from "monaco-editor";
 import * as Constants from "./constants";
 import * as Designer from "./card-designer-surface";
 import * as DesignerPeers from "./designer-peers";
+import { SamplePickerDialog } from "./sample-picker-dialog";
 import { HostContainer } from "./containers/host-container";
 import { adaptiveCardSchema } from "./adaptive-card-schema";
 import { FullScreenHandler } from "./fullscreen-handler";
 import { Toolbar, ToolbarButton, ToolbarChoicePicker, ToolbarElementAlignment } from "./toolbar";
-import { SidePane, SidePaneOrientation } from "./side-pane";
-import { Splitter } from "./splitter";
-import { IPoint, Utils } from "./miscellaneous";
-import { BasePaletteItem, ElementPaletteItem } from "./tool-palette";
+import { IPoint, Utils, defaultHostConfig } from "./miscellaneous";
+import { BasePaletteItem, ElementPaletteItem, DataPaletteItem, SnippetPaletteItem, CustomPaletteItem } from "./tool-palette";
 import { DefaultContainer } from "./containers/default/default-container";
+import { SidePanel, SidePanelAlignment } from "./side-panel";
+import { Toolbox, IToolboxCommand } from "./tool-box";
+import { FieldDefinition } from "./data";
+import { DataTreeItem } from "./data-treeitem";
+import { BaseTreeItem } from "./base-tree-item";
+import * as Shared from "./shared";
 
 export class CardDesigner {
     private static internalProcessMarkdown(text: string, result: Adaptive.IMarkdownProcessingResult) {
         if (CardDesigner.onProcessMarkdown) {
             CardDesigner.onProcessMarkdown(text, result);
-        }
-        else {
+        } else {
             // Check for markdownit
             if (window["markdownit"]) {
                 result.outputHtml = window["markdownit"]().render(text);
@@ -34,11 +38,12 @@ export class CardDesigner {
 
     private static MAX_UNDO_STACK_SIZE = 50;
 
-    private _monacoEditor: monaco.editor.IStandaloneCodeEditor;
+    private _isAttached: boolean = false;
+    private _cardEditor: monaco.editor.IStandaloneCodeEditor;
+    private _sampleDataEditor: monaco.editor.IStandaloneCodeEditor;
     private _hostContainers: Array<HostContainer>;
     private _isMonacoEditorLoaded: boolean = false;
     private _designerSurface: Designer.CardDesignerSurface;
-    private _propertySheetHostConfig: Adaptive.HostConfig;
     private _designerHostElement: HTMLElement;
     private _draggedPaletteItem: BasePaletteItem;
     private _draggedElement: HTMLElement;
@@ -47,30 +52,91 @@ export class CardDesigner {
     private _activeHostContainer: HostContainer;
     private _undoStack: Array<object> = [];
     private _undoStackIndex: number = -1;
-    private _toolPalettePane: SidePane;
-    private _jsonEditorPane: SidePane;
-    private _propertySheetPane: SidePane;
-	private _treeViewPane: SidePane;
-	private _assetPath: string;
+    private _toolPaletteToolbox: Toolbox;
+    private _propertySheetToolbox: Toolbox;
+    private _treeViewToolbox: Toolbox;
+    private _jsonEditorsPanel: SidePanel;
+    private _cardEditorToolbox: Toolbox;
+    private _sampleDataEditorToolbox: Toolbox;
+    private _dataToolbox: Toolbox;
+    private _assetPath: string;
+    private _dataStructure: FieldDefinition;
+    private _sampleData: any;
+    private _customPeletteItems: CustomPaletteItem[];
+
+    private togglePreview() {
+        this._designerSurface.isPreviewMode = !this._designerSurface.isPreviewMode;
+
+        if (this._designerSurface.isPreviewMode) {
+            this._togglePreviewButton.toolTip = "Return to Design mode";
+        }
+        else {
+            this._togglePreviewButton.toolTip = "Switch to Preview mode";
+            this._designerSurface.updateLayout();
+        }
+
+        this.buildTreeView();
+    }
 
     private buildTreeView() {
-        if (this._treeViewPane.content) {
-            this._treeViewPane.content.innerHTML = "";
-            this._treeViewPane.content.appendChild(this.designerSurface.rootPeer.treeItem.render());
-		}
+        if (this._treeViewToolbox.content) {
+            this._treeViewToolbox.content.innerHTML = "";
+
+            if (this.designerSurface.isPreviewMode) {
+                this.treeViewToolbox.content.innerHTML =
+                    '<div style="padding: 8px; display: flex; justify-content: center;">' +
+                        '<div>The Card Structure isn\'t available in Preview mode.</div>' +
+                    '</div>';
+            }
+            else {
+                this._treeViewToolbox.content.appendChild(this.designerSurface.rootPeer.treeItem.render());
+            }
+        }
+    }
+
+    private setupDataTreeItemEvents(treeItem: DataTreeItem) {
+        treeItem.onStartDrag = (sender: BaseTreeItem) => {
+            this._draggedPaletteItem = new DataPaletteItem(treeItem.field);
+
+            this._draggedElement = this._draggedPaletteItem.renderDragVisual();
+            this._draggedElement.style.position = "absolute";
+            this._draggedElement.style.left = this._currentMousePosition.x + "px";
+            this._draggedElement.style.top = this._currentMousePosition.y + "px";
+
+            document.body.appendChild(this._draggedElement);
+
+            treeItem.endDrag();
+        }
+
+        for (let i = 0; i < treeItem.getChildCount(); i++) {
+            this.setupDataTreeItemEvents(treeItem.getChildAt(i));
+        }
+    }
+
+    private buildDataExplorer() {
+        if (this._dataToolbox && this._dataToolbox.content) {
+            this._dataToolbox.content.innerHTML = "";
+
+            if (this._dataStructure) {
+                let treeItem = new DataTreeItem(this._dataStructure);
+
+                this._dataToolbox.content.appendChild(treeItem.render());
+
+                this.setupDataTreeItemEvents(treeItem);
+            }
+        }
     }
 
     private buildPropertySheet(peer: DesignerPeers.DesignerPeer) {
-        if (this._propertySheetPane.content) {
-            this._propertySheetPane.content.innerHTML = "";
+        if (this._propertySheetToolbox.content) {
+            this._propertySheetToolbox.content.innerHTML = "";
 
             let card: Adaptive.AdaptiveCard;
 
             if (peer) {
-                card = peer.buildPropertySheetCard();
+                card = peer.buildPropertySheetCard(this.currentTargetVersion);
             }
             else {
-
                 card = new Adaptive.AdaptiveCard();
                 card.parse(
                     {
@@ -98,9 +164,10 @@ export class CardDesigner {
                 )
             }
 
-            card.hostConfig = this._propertySheetHostConfig;
+            // card.hostConfig = this._propertySheetHostConfig;
+            card.hostConfig = defaultHostConfig;
 
-            this._propertySheetPane.content.appendChild(card.render());
+            this._propertySheetToolbox.content.appendChild(card.render());
         }
     }
 
@@ -109,7 +176,7 @@ export class CardDesigner {
         paletteItem.onStartDrag = (sender: BasePaletteItem) => {
             this._draggedPaletteItem = sender;
 
-            this._draggedElement = sender.cloneElement();
+            this._draggedElement = sender.renderDragVisual();
             this._draggedElement.style.position = "absolute";
             this._draggedElement.style.left = this._currentMousePosition.x + "px";
             this._draggedElement.style.top = this._currentMousePosition.y + "px";
@@ -121,7 +188,11 @@ export class CardDesigner {
     }
 
     private buildPalette() {
-        this._toolPalettePane.content.innerHTML = "";
+        if (!this._isAttached) {
+            return;
+        }
+
+        this._toolPaletteToolbox.content.innerHTML = "";
 
         let categorizedTypes: Object = {};
 
@@ -143,56 +214,27 @@ export class CardDesigner {
             }
         }
 
+        if (this.customPaletteItems) {
+            for (let item of this.customPaletteItems) {
+                if (!categorizedTypes.hasOwnProperty(item.category)) {
+                    categorizedTypes[item.category] = [];
+                }
+
+                categorizedTypes[item.category].push(item);
+            }
+        }
+
         for (let category in categorizedTypes) {
             let node = document.createElement('li');
             node.innerText = category;
             node.className = "acd-palette-category";
 
-            this._toolPalettePane.content.appendChild(node);
+            this._toolPaletteToolbox.content.appendChild(node);
 
             for (var i = 0; i < categorizedTypes[category].length; i++) {
-                this.addPaletteItem(categorizedTypes[category][i], this._toolPalettePane.content);
+                this.addPaletteItem(categorizedTypes[category][i], this._toolPaletteToolbox.content);
             }
         }
-
-        /* This is to test "snippet" support. Snippets are not yet fully baked
-        let personaHeaderSnippet = new SnippetPaletteItem("Persona header");
-        personaHeaderSnippet.snippet = {
-            type: "ColumnSet",
-            columns: [
-                {
-                    width: "auto",
-                    items: [
-                        {
-                            type: "Image",
-                            size: "Small",
-                            style: "Person",
-                            url: "https://pbs.twimg.com/profile_images/3647943215/d7f12830b3c17a5a9e4afcc370e3a37e_400x400.jpeg"
-                        }
-                    ]
-                },
-                {
-                    width: "stretch",
-                    items: [
-                        {
-                            type: "TextBlock",
-                            text: "John Doe",
-                            weight: "Bolder",
-                            wrap: true
-                        },
-                        {
-                            type: "TextBlock",
-                            spacing: "None",
-                            text: "Additional information",
-                            wrap: true
-                        }
-                    ]
-                }
-            ]
-        };
-
-        this.addPaletteItem(personaHeaderSnippet);
-        */
     }
 
     private endDrag() {
@@ -249,16 +291,18 @@ export class CardDesigner {
 			styleSheetLinkElement.href = Utils.joinPaths(this._assetPath, this.activeHostContainer.styleSheet);
 		}
 
-        let designerBackground = document.getElementById("designerBackground");
+        let _cardArea = document.getElementById("cardArea");
 
-        if (designerBackground) {
-            designerBackground.style.backgroundColor = this.activeHostContainer.getBackgroundColor();
+        if (_cardArea) {
+            _cardArea.style.backgroundColor = this.activeHostContainer.getBackgroundColor();
         }
 
         this.activeHostContainer.initialize();
 
         this._designerHostElement.innerHTML = "";
         this.activeHostContainer.renderTo(this._designerHostElement);
+
+        let wasInPreviewMode = this._designerSurface ? this._designerSurface.isPreviewMode : false;
 
         this._designerSurface = new Designer.CardDesignerSurface(this.activeHostContainer.cardHost);
         this._designerSurface.fixedHeightCard = this.activeHostContainer.isFixedHeight;
@@ -314,28 +358,47 @@ export class CardDesigner {
         }
 
         this._designerSurface.card = this._card;
+
+        this.updateSampleData();
+
+        this._designerSurface.isPreviewMode = wasInPreviewMode;
+
+        this.updateFullLayout();
     }
 
     private activeHostContainerChanged() {
         this.recreateDesignerSurface();
+
+        if (this.onActiveHostContainerChanged) {
+            this.onActiveHostContainerChanged(this);
+        }
     }
 
-    public updateJsonEditorLayout() {
+    private updateToolboxLayout(toolbox: Toolbox, hostPanelRect: ClientRect | DOMRect) {
+        if (toolbox) {
+            let jsonEditorHeaderRect = toolbox.getHeaderBoundingRect();
+
+            toolbox.content.style.height = (hostPanelRect.height - jsonEditorHeaderRect.height) + "px";
+        }
+    }
+
+    public updateJsonEditorsLayout() {
         if (this._isMonacoEditorLoaded) {
-            // Monaco is very finicky. It will apparently only properly layout if
-            // its direct container has an explicit height.
-            let jsonEditorPaneRect = this._jsonEditorPane.attachedTo.getBoundingClientRect();
-            let jsonEditorHeaderRect = this._jsonEditorPane.getHeaderBoundingRect();
+            let jsonEditorsPaneRect = this._jsonEditorsPanel.contentHost.getBoundingClientRect();
 
-            this._jsonEditorPane.content.style.height = (jsonEditorPaneRect.height - jsonEditorHeaderRect.height) + "px";
+            this.updateToolboxLayout(this._cardEditorToolbox, jsonEditorsPaneRect);
+            this._cardEditor.layout();
 
-            this._monacoEditor.layout();
+            if (this._sampleDataEditorToolbox) {
+                this.updateToolboxLayout(this._sampleDataEditorToolbox, jsonEditorsPaneRect);
+                this._sampleDataEditor.layout();
+            }
         }
     }
 
     private updateFullLayout() {
         this.scheduleLayoutUpdate();
-        this.updateJsonEditorLayout();
+        this.updateJsonEditorsLayout();
     }
 
     private jsonUpdateTimer: any;
@@ -344,8 +407,24 @@ export class CardDesigner {
 
     private preventCardUpdate: boolean = false;
 
-    private setJsonPayload(payload: object) {
-        this._monacoEditor.setValue(JSON.stringify(payload, null, 4));
+    private cardPayloadChanged() {
+        if (this.onCardPayloadChanged) {
+            this.onCardPayloadChanged(this);
+        }
+    }
+    
+    private setCardPayload(payload: object) {
+        if (this._isMonacoEditorLoaded) {
+            this._cardEditor.setValue(JSON.stringify(payload, null, 4));
+        }
+
+        this.cardPayloadChanged();
+    }
+
+    private setSampleDataPayload(payload: any) {
+        if (this._isMonacoEditorLoaded && this._sampleDataEditor) {
+            this._sampleDataEditor.setValue(JSON.stringify(payload, null, 4));
+        }
     }
 
     private updateJsonFromCard(addToUndoStack: boolean = true) {
@@ -358,8 +437,8 @@ export class CardDesigner {
                 if (addToUndoStack) {
                     this.addToUndoStack(cardPayload);
                 }
-
-                this.setJsonPayload(cardPayload);
+    
+                this.setCardPayload(cardPayload);
             }
         }
         finally {
@@ -376,9 +455,13 @@ export class CardDesigner {
     }
 
     private preventJsonUpdate: boolean = false;
+    
+    private getCurrentCardEditorPayload(): string {
+        return this._isMonacoEditorLoaded ? this._cardEditor.getValue() : Constants.defaultPayload;
+    }
 
-    private getCurrentJsonPayload(): string {
-        return this._isMonacoEditorLoaded ? this._monacoEditor.getValue() : Constants.defaultPayload;
+    private getCurrentSampleDataEditorPayload(): string {
+        return this._isMonacoEditorLoaded && this._sampleDataEditor ? this._sampleDataEditor.getValue() : "";
     }
 
     private updateCardFromJson() {
@@ -386,7 +469,9 @@ export class CardDesigner {
             this.preventJsonUpdate = true;
 
             if (!this.preventCardUpdate) {
-                this.designerSurface.setCardPayloadAsString(this.getCurrentJsonPayload());
+                this.designerSurface.setCardPayloadAsString(this.getCurrentCardEditorPayload());
+
+                this.cardPayloadChanged();
             }
         }
         finally {
@@ -411,12 +496,32 @@ export class CardDesigner {
     private _fullScreenHandler = new FullScreenHandler();
     private _fullScreenButton: ToolbarButton;
     private _hostContainerChoicePicker: ToolbarChoicePicker;
+    private _versionChoicePicker: ToolbarChoicePicker;
     private _undoButton: ToolbarButton;
     private _redoButton: ToolbarButton;
     private _newCardButton: ToolbarButton;
     private _copyJSONButton: ToolbarButton;
+    private _togglePreviewButton: ToolbarButton;
 
     private prepareToolbar() {
+        if (Shared.GlobalSettings.previewFeaturesEnabled) {
+            this._versionChoicePicker = new ToolbarChoicePicker(CardDesigner.ToolbarCommands.VersionPicker);
+            this._versionChoicePicker.label = "Target version:"
+            this._versionChoicePicker.width = 80;
+            this._versionChoicePicker.alignment = ToolbarElementAlignment.Right;
+            this._versionChoicePicker.separator = true;
+
+            for (let i = 0; i < Shared.SupportedTargetVersions.length; i++) {
+                this._versionChoicePicker.choices.push(
+                    {
+                        name: Shared.SupportedTargetVersions[i].label,
+                        value: i.toString()
+                    });
+            }
+
+            this.toolbar.addElement(this._versionChoicePicker);
+        }
+
         this._fullScreenButton = new ToolbarButton(
             CardDesigner.ToolbarCommands.FullScreen,
             "Enter Full Screen",
@@ -427,6 +532,49 @@ export class CardDesigner {
         this._fullScreenButton.alignment = ToolbarElementAlignment.Right;
 
         this.toolbar.addElement(this._fullScreenButton);
+
+        let openSampleButton = new ToolbarButton(
+            CardDesigner.ToolbarCommands.OpenPayload,
+            "Open Sample",
+            "acd-icon-open",
+            (sender: ToolbarButton) => {
+                let dialog = new SamplePickerDialog();
+                dialog.title = "Pick a sample";
+                dialog.width = "30%";
+                dialog.height = "60%";
+                dialog.catalogueUrl = this.sampleCatalogueUrl;
+                dialog.onClose = (d) => {
+                    if (dialog.selectedSample) {
+                        dialog.selectedSample.onDownloaded = () => {
+                            try {
+                                let cardPayload = JSON.parse(dialog.selectedSample.cardPayload);
+
+                                this.setCardPayload(cardPayload);
+                            }
+                            catch {
+                                alert("The sample could not be loaded.")
+                            }
+
+                            if (!Adaptive.isNullOrEmpty(dialog.selectedSample.sampleData)) {
+                                try {
+                                    let sampleDataPayload = JSON.parse(dialog.selectedSample.sampleData);
+    
+                                    this.setSampleDataPayload(sampleDataPayload);
+                                    this.dataStructure = FieldDefinition.create(sampleDataPayload);
+                                }
+                                catch {
+                                    alert("The sample could not be loaded.")
+                                }
+                            }
+                        };
+                        dialog.selectedSample.download();
+                    }
+                };
+                dialog.open();
+            }
+        )
+
+        this.toolbar.addElement(openSampleButton);
 
         if (this._hostContainers && this._hostContainers.length > 0) {
             this._hostContainerChoicePicker = new ToolbarChoicePicker(CardDesigner.ToolbarCommands.HostAppPicker);
@@ -456,7 +604,7 @@ export class CardDesigner {
             CardDesigner.ToolbarCommands.Undo,
             "Undo",
             "acd-icon-undo",
-            (sender) => { this.undo(); });
+            (sender: ToolbarButton) => { this.undo(); });
         this._undoButton.separator = true;
         this._undoButton.toolTip = "Undo your last change";
         this._undoButton.isEnabled = false;
@@ -468,7 +616,7 @@ export class CardDesigner {
             CardDesigner.ToolbarCommands.Redo,
             "Redo",
             "acd-icon-redo",
-            (sender) => { this.redo(); });
+            (sender: ToolbarButton) => { this.redo(); });
         this._redoButton.toolTip = "Redo your last changes";
         this._redoButton.isEnabled = false;
         this._redoButton.displayCaption = false;
@@ -479,16 +627,25 @@ export class CardDesigner {
             CardDesigner.ToolbarCommands.NewCard,
             "New card",
             "acd-icon-newCard",
-            (sender) => { this.newCard(); });
-            this._newCardButton.separator = true;
+            (sender: ToolbarButton) => { this.newCard(); });
+        this._newCardButton.separator = true;
 
         this.toolbar.addElement(this._newCardButton);
 
         this._copyJSONButton = new ToolbarButton(
             CardDesigner.ToolbarCommands.CopyJSON,
-            "Copy JSON",
+            "Copy card JSON",
             "acd-icon-copy");
         this.toolbar.addElement(this._copyJSONButton);
+
+        this._togglePreviewButton = new ToolbarButton(
+            CardDesigner.ToolbarCommands.TogglePreview,
+            "Preview mode",
+            "acd-icon-preview",
+            (sender: ToolbarButton) => { this.togglePreview(); });
+        this._togglePreviewButton.separator = true;
+        this._togglePreviewButton.allowToggle = true;
+        this.toolbar.addElement(this._togglePreviewButton);
 
         this._fullScreenHandler = new FullScreenHandler();
         this._fullScreenHandler.onFullScreenChanged = (isFullScreen: boolean) => {
@@ -499,19 +656,21 @@ export class CardDesigner {
     }
 
     private onResize() {
-        this._monacoEditor.layout();
+        this._cardEditor.layout();
+
+        if (this._sampleDataEditor) {
+            this._sampleDataEditor.layout();
+        }
     }
 
-    private loadMonaco(callback: () => void) {
-        // window["require"].config({ paths: { 'vs': './editor/monaco/min/vs' } });
-        // window["require"](
-        //     ['vs/editor/editor.main'],
-        //     function () {
-        //         callback();
-		//     });
-
-		// If loaded using WebPack this should work, but it's not right now...
-		//callback();
+    private updateSampleData() {
+        try {
+            this._sampleData = JSON.parse(this.getCurrentSampleDataEditorPayload());
+            this.designerSurface.sampleData = this._sampleData;
+        }
+        catch {
+            // Swallow expression, the payload isn't a valid JSON document
+        }
     }
 
     public monacoModuleLoaded(monaco: any = null) {
@@ -530,18 +689,18 @@ export class CardDesigner {
             validate: false,
             allowComments: true
         }
-
-		this._jsonEditorPane.content = document.createElement("div");
-		this._jsonEditorPane.content.style.overflow = "hidden";
-
+    
 		// TODO: set this in our editor instead of defaults
         monaco.languages.json.jsonDefaults.setDiagnosticsOptions(monacoConfiguration);
 
-        this._monacoEditor = monaco.editor.create(
-            this._jsonEditorPane.content,
+        // Setup card JSON editor
+        this._cardEditorToolbox.content = document.createElement("div");
+        this._cardEditorToolbox.content.style.overflow = "hidden";
+		
+        this._cardEditor = monaco.editor.create(
+            this._cardEditorToolbox.content,
             {
                 folding: true,
-                // validate: false,
                 fontSize: 13.5,
                 language: 'json',
                 minimap: {
@@ -549,14 +708,37 @@ export class CardDesigner {
                 }
             }
         );
+        
+        this._cardEditor.onDidChangeModelContent(() => { this.scheduleUpdateCardFromJson(); });
 
-        this._monacoEditor.onDidChangeModelContent(() => { this.scheduleUpdateCardFromJson(); });
+        if (this._sampleDataEditorToolbox) {
+            // Setup sample data JSON editor
+            this._sampleDataEditorToolbox.content = document.createElement("div");
+            this._sampleDataEditorToolbox.content.style.overflow = "hidden";
+            
+            this._sampleDataEditor = monaco.editor.create(
+                this._sampleDataEditorToolbox.content,
+                {
+                    folding: true,
+                    fontSize: 13.5,
+                    language: 'json',
+                    minimap: {
+                        enabled: false
+                    }
+                }
+            );
+            
+            this._sampleDataEditor.onDidChangeModelContent(
+                () => {
+                    this.updateSampleData();
+                });
+        }
 
         window.addEventListener('resize', () => { this.onResize(); });
 
         this._isMonacoEditorLoaded = true;
 
-        this.updateJsonEditorLayout();
+        this.updateJsonEditorsLayout();
         this.updateJsonFromCard(false);
     }
 
@@ -624,7 +806,10 @@ export class CardDesigner {
 
     readonly toolbar: Toolbar = new Toolbar();
 
+    sampleCatalogueUrl: string = undefined;
+
     constructor(hostContainers: Array<HostContainer> = null) {
+        Adaptive.AdaptiveCard.enableFullJsonRoundTrip = true;
         Adaptive.AdaptiveCard.onProcessMarkdown = (text: string, result: Adaptive.IMarkdownProcessingResult) => {
             CardDesigner.internalProcessMarkdown(text, result);
         }
@@ -632,287 +817,6 @@ export class CardDesigner {
         this._hostContainers = hostContainers ? hostContainers : [];
 
         this.prepareToolbar();
-
-        this._propertySheetHostConfig = new Adaptive.HostConfig(
-            {
-                preExpandSingleShowCardAction: true,
-                supportsInteractivity: true,
-                spacing: {
-                    small: 10,
-                    default: 20,
-                    medium: 30,
-                    large: 40,
-                    extraLarge: 50,
-                    padding: 20
-                },
-                separator: {
-                    lineThickness: 1,
-                    lineColor: "#EEEEEE"
-                },
-                textAlign: {
-                    right: "right"
-                },
-                imageSizes: {
-                    small: 40,
-                    medium: 80,
-                    large: 160
-                },
-                fontTypes: {
-                    default: {
-                        fontFamily: "'Segoe UI', Tahoma, Geneva, Verdana, sans-serif",
-                        fontSizes: {
-                            small: 12,
-                            default: 14,
-                            medium: 17,
-                            large: 21,
-                            extraLarge: 26
-                        },
-                        fontWeights: {
-                            lighter: 200,
-                            default: 400,
-                            bolder: 600
-                        }
-                    },
-                    monospace: {
-                        fontFamily: "'Courier New', Courier, monospace",
-                        fontSizes: {
-                            small: 12,
-                            default: 14,
-                            medium: 17,
-                            large: 21,
-                            extraLarge: 26
-                        },
-                        fontWeights: {
-                            lighter: 200,
-                            default: 400,
-                            bolder: 600
-                        }
-                    }
-                },
-                containerStyles: {
-                    default: {
-                        backgroundColor: "#f9f9f9",
-                        foregroundColors: {
-                            default: {
-                                default: "#333333",
-                                subtle: "#EE333333"
-                            },
-                            accent: {
-                                default: "#2E89FC",
-                                subtle: "#882E89FC"
-                            },
-                            attention: {
-                                default: "#cc3300",
-                                subtle: "#DDcc3300"
-                            },
-                            good: {
-                                default: "#54a254",
-                                subtle: "#DD54a254"
-                            },
-                            warning: {
-                                default: "#e69500",
-                                subtle: "#DDe69500"
-                            }
-                        }
-                    },
-                    emphasis: {
-                        backgroundColor: "#08000000",
-                        foregroundColors: {
-                            default: {
-                                default: "#333333",
-                                subtle: "#EE333333"
-                            },
-                            accent: {
-                                default: "#2E89FC",
-                                subtle: "#882E89FC"
-                            },
-                            attention: {
-                                default: "#cc3300",
-                                subtle: "#DDcc3300"
-                            },
-                            good: {
-                                default: "#54a254",
-                                subtle: "#DD54a254"
-                            },
-                            warning: {
-                                default: "#e69500",
-                                subtle: "#DDe69500"
-                            }
-                        }
-                    },
-                    accent: {
-                        backgroundColor: "#C7DEF9",
-                        foregroundColors: {
-                            default: {
-                                default: "#333333",
-                                subtle: "#EE333333"
-                            },
-                            dark: {
-                                default: "#000000",
-                                subtle: "#66000000"
-                            },
-                            light: {
-                                default: "#FFFFFF",
-                                subtle: "#33000000"
-                            },
-                            accent: {
-                                default: "#2E89FC",
-                                subtle: "#882E89FC"
-                            },
-                            attention: {
-                                default: "#cc3300",
-                                subtle: "#DDcc3300"
-                            },
-                            good: {
-                                default: "#54a254",
-                                subtle: "#DD54a254"
-                            },
-                            warning: {
-                                default: "#e69500",
-                                subtle: "#DDe69500"
-                            }
-                        }
-                    },
-                    good: {
-                        backgroundColor: "#CCFFCC",
-                        foregroundColors: {
-                            default: {
-                                default: "#333333",
-                                subtle: "#EE333333"
-                            },
-                            dark: {
-                                default: "#000000",
-                                subtle: "#66000000"
-                            },
-                            light: {
-                                default: "#FFFFFF",
-                                subtle: "#33000000"
-                            },
-                            accent: {
-                                default: "#2E89FC",
-                                subtle: "#882E89FC"
-                            },
-                            attention: {
-                                default: "#cc3300",
-                                subtle: "#DDcc3300"
-                            },
-                            good: {
-                                default: "#54a254",
-                                subtle: "#DD54a254"
-                            },
-                            warning: {
-                                default: "#e69500",
-                                subtle: "#DDe69500"
-                            }
-                        }
-                    },
-                    attention: {
-                        backgroundColor: "#FFC5B2",
-                        foregroundColors: {
-                            default: {
-                                default: "#333333",
-                                subtle: "#EE333333"
-                            },
-                            dark: {
-                                default: "#000000",
-                                subtle: "#66000000"
-                            },
-                            light: {
-                                default: "#FFFFFF",
-                                subtle: "#33000000"
-                            },
-                            accent: {
-                                default: "#2E89FC",
-                                subtle: "#882E89FC"
-                            },
-                            attention: {
-                                default: "#cc3300",
-                                subtle: "#DDcc3300"
-                            },
-                            good: {
-                                default: "#54a254",
-                                subtle: "#DD54a254"
-                            },
-                            warning: {
-                                default: "#e69500",
-                                subtle: "#DDe69500"
-                            }
-                        }
-                    },
-                    warning: {
-                        backgroundColor: "#FFE2B2",
-                        foregroundColors: {
-                            default: {
-                                default: "#333333",
-                                subtle: "#EE333333"
-                            },
-                            dark: {
-                                default: "#000000",
-                                subtle: "#66000000"
-                            },
-                            light: {
-                                default: "#FFFFFF",
-                                subtle: "#33000000"
-                            },
-                            accent: {
-                                default: "#2E89FC",
-                                subtle: "#882E89FC"
-                            },
-                            attention: {
-                                default: "#cc3300",
-                                subtle: "#DDcc3300"
-                            },
-                            good: {
-                                default: "#54a254",
-                                subtle: "#DD54a254"
-                            },
-                            warning: {
-                                default: "#e69500",
-                                subtle: "#DDe69500"
-                            }
-                        }
-                    }
-                },
-                actions: {
-                    maxActions: 5,
-                    spacing: Adaptive.Spacing.Default,
-                    buttonSpacing: 10,
-                    showCard: {
-                        actionMode: Adaptive.ShowCardActionMode.Inline,
-                        inlineTopMargin: 16
-                    },
-                    actionsOrientation: Adaptive.Orientation.Horizontal,
-                    actionAlignment: Adaptive.ActionAlignment.Left
-                },
-                adaptiveCard: {
-                    allowCustomStyle: true
-                },
-                imageSet: {
-                    imageSize: Adaptive.Size.Medium,
-                    maxImageHeight: 100
-                },
-                factSet: {
-                    title: {
-                        color: Adaptive.TextColor.Default,
-                        size: Adaptive.TextSize.Default,
-                        isSubtle: false,
-                        weight: Adaptive.TextWeight.Bolder,
-                        wrap: true,
-                        maxWidth: 150,
-                    },
-                    value: {
-                        color: Adaptive.TextColor.Default,
-                        size: Adaptive.TextSize.Default,
-                        isSubtle: false,
-                        weight: Adaptive.TextWeight.Default,
-                        wrap: true,
-                    },
-                    spacing: 10
-                }
-            }
-        );
-
-        this._propertySheetHostConfig.cssClassNamePrefix = "default";
     }
 
     attachTo(root: HTMLElement)  {
@@ -928,7 +832,7 @@ export class CardDesigner {
             this._activeHostContainer = this._hostContainers[0];
         }
         else {
-            this._activeHostContainer = new DefaultContainer("Default", "default-container.css");
+            this._activeHostContainer = new DefaultContainer("Default", "adaptivecards-defaulthost.css");
         }
 
         root.style.flex = "1 1 auto";
@@ -939,139 +843,155 @@ export class CardDesigner {
         root.innerHTML =
             '<div id="toolbarHost"></div>' +
             '<div class="content" style="display: flex; flex: 1 1 auto; overflow-y: hidden;">' +
-                '<div id="leftCollapsedPaneTabHost" class="acd-verticalCollapsedTabContainer" style="border-right: 1px solid #D2D2D2;"></div>' +
-                '<div id="toolPalettePane" class="selector-toolPalette" style="display: flex; flex-direction: column; background-color: white; border-right: 1px solid #D2D2D2;">' +
-                    '<div id="toolPaletteHost" class="acd-dockedPane" style="overflow: auto"></div>' +
-                '</div>' +
+                '<div id="leftCollapsedPaneTabHost" class="acd-verticalCollapsedTabContainer acd-dockedLeft" style="border-right: 1px solid #D2D2D2;"></div>' +
+                '<div id="toolPalettePanel" class="acd-toolPalette-pane"></div>' +
                 '<div style="display: flex; flex-direction: column; flex: 1 1 100%; overflow: hidden;">' +
                     '<div style="display: flex; flex: 1 1 100%; overflow: hidden;">' +
-                        '<div id="designerBackground" style="flex: 1 1 70%; background-color: #F6F6F6; display: flex; flex-direction: column; overflow: auto;">' +
+                        '<div id="cardArea" class="acd-designer-cardArea">' +
                             '<div style="flex: 1 1 100%; overflow: auto;">' +
                                 '<div id="designerHost" style="margin: 20px 40px 20px 20px;"></div>' +
                             '</div>' +
                             '<div id="errorPane" class="acd-error-pane acd-hidden"></div>' +
                         '</div>' +
-                        '<div id="treeViewSplitter" class="acd-vertical-splitter selector-treeView"></div>' +
-                        '<div id="treeViewPane" class="acd-treeView-pane selector-treeView">' +
-                            '<div id="treeViewHost" class="acd-treeView-host"></div>' +
-                        '</div>' +
-                        '<div id="propertySheetSplitter" class="acd-vertical-splitter selector-propertySheet"></div>' +
-                        '<div id="propertySheetPane" class="acd-propertySheet-pane selector-propertySheet">' +
-                            '<div id="propertySheetHost" class="acd-propertySheet-host"></div>' +
-                        '</div>' +
+                        '<div id="treeViewPanel" class="acd-treeView-pane"></div>' +
+                       '<div id="propertySheetPanel" class="acd-propertySheet-pane"></div>' +
                     '</div>' +
-                    '<div id="jsonEditorSplitter" class="acd-horizontal-splitter selector-jsonEditor"></div>' +
-                    '<div id="jsonEditorPane" class="acd-json-editor-pane selector-jsonEditor"></div>' +
-                    '<div id="bottomCollapsedPaneTabHost" style="border-top: 1px solid #D2D2D2;"></div>' +
+                    '<div id="jsonEditorPanel" class="acd-json-editor-pane"></div>' +
+                    '<div id="bottomCollapsedPaneTabHost" class="acd-horizontalCollapsedTabContainer" style="border-top: 1px solid #D2D2D2;"></div>' +
                 '</div>' +
-                '<div id="rightCollapsedPaneTabHost" class="acd-verticalCollapsedTabContainer" style="border-left: 1px solid #D2D2D2;"></div>' +
+                '<div id="rightCollapsedPaneTabHost" class="acd-verticalCollapsedTabContainer acd-dockedRight" style="border-left: 1px solid #D2D2D2;"></div>' +
             '</div>';
 
         this.toolbar.attachTo(document.getElementById("toolbarHost"));
 
-        new Clipboard(
-            this._copyJSONButton.renderedElement,
-            {
-                text: (trigger) => { return JSON.stringify(this.card.toJSON(), null, 4); }
-            });
-
-        // Tool palette pane
-        this._toolPalettePane = new SidePane(
-            document.getElementById("toolPalettePane"),
-            document.getElementById("leftCollapsedPaneTabHost"),
-            "toolPalette",
-            "Tool box",
-            "selector-toolPalette");
-        this._toolPalettePane.onToggled = (sender: SidePane) => {
-            this.updateFullLayout();
-        }
-        this._toolPalettePane.content = document.getElementById("toolPaletteHost");
-
-        // Splitter for JSON editor pane
-        let jsonEditorSplitter = new Splitter(
-            document.getElementById("jsonEditorSplitter"),
-            document.getElementById("jsonEditorPane"),
-            false);
-        jsonEditorSplitter.onResized = (splitter: Splitter, newSize: number) => {
-            this.updateJsonEditorLayout();
-
-            this._jsonEditorPane.saveState();
+        if (this._versionChoicePicker) {
+            this._versionChoicePicker.selectedIndex = Shared.SupportedTargetVersions.indexOf(Shared.Versions.v1_2);
+            this._versionChoicePicker.onChanged = (sender) => {
+                this.buildPropertySheet(this._designerSurface.selectedPeer);
+            }
         }
 
-        // JSON editor pane
-        this._jsonEditorPane = new SidePane(
-            document.getElementById("jsonEditorPane"),
-            document.getElementById("bottomCollapsedPaneTabHost"),
-            "jsonEditor",
-            "JSON",
-            "selector-jsonEditor",
-            SidePaneOrientation.Horizontal
-        )
-        this._jsonEditorPane.onToggled = (sender: SidePane) => {
-            this.updateFullLayout();
+        if (this._copyJSONButton.isVisible) {
+            new Clipboard(
+                this._copyJSONButton.renderedElement,
+                {
+                    text: (trigger) => { return JSON.stringify(this.card.toJSON(), null, 4); }
+                });
         }
-        this._jsonEditorPane.content = document.createElement("div");
-        this._jsonEditorPane.content.style.padding = "8px";
-        this._jsonEditorPane.content.innerText = "Loading editor..."
+        
+        // Tool palette panel
+        let toolPaletteHost = document.createElement("div");
+        toolPaletteHost.className = "acd-dockedPane";
 
-        // Splitter for property sheet pane
-        let propertySheetSplitter = new Splitter(
-            document.getElementById("propertySheetSplitter"),
-            document.getElementById("propertySheetPane"),
-            true);
-        propertySheetSplitter.minimum = 230;
-        propertySheetSplitter.onResized = (splitter: Splitter, newSize: number) => {
+        this._toolPaletteToolbox = new Toolbox("toolPalette", "Card Elements");
+        this._toolPaletteToolbox.content = toolPaletteHost;
+
+        let toolPalettePanel = new SidePanel(
+            "toolPalettePanel",
+            SidePanelAlignment.Left,
+            document.getElementById("leftCollapsedPaneTabHost"));
+        toolPalettePanel.addToolbox(this._toolPaletteToolbox);
+        toolPalettePanel.isResizable = false;
+
+        toolPalettePanel.attachTo(document.getElementById("toolPalettePanel"));
+
+        // JSON editors panel
+        this._cardEditorToolbox = new Toolbox("cardEditor", "Card Payload Editor");
+        this._cardEditorToolbox.content = document.createElement("div");
+        this._cardEditorToolbox.content.style.padding = "8px";
+        this._cardEditorToolbox.content.innerText = "Loading editor...";
+
+        this._jsonEditorsPanel = new SidePanel(
+            "jsonEditorPanel",
+            SidePanelAlignment.Bottom,
+            document.getElementById("bottomCollapsedPaneTabHost"));
+        this._jsonEditorsPanel.onResized = (sender: SidePanel) => {
+            this.updateJsonEditorsLayout();
+        }
+        this._jsonEditorsPanel.onToolboxResized = (sender: SidePanel) => {
+            this.updateJsonEditorsLayout();
+        }
+        this._jsonEditorsPanel.onToolboxExpandedOrCollapsed = (sender: SidePanel) => {
+            this.updateJsonEditorsLayout();
+        }
+
+        this._jsonEditorsPanel.addToolbox(this._cardEditorToolbox);
+
+        if (Shared.GlobalSettings.previewFeaturesEnabled) {
+            this._sampleDataEditorToolbox = new Toolbox("sampleDataEditor", "Sample Data Editor");
+            this._sampleDataEditorToolbox.content = document.createElement("div");
+            this._sampleDataEditorToolbox.content.style.padding = "8px";
+            this._sampleDataEditorToolbox.content.innerText = "Loading editor...";
+            this._sampleDataEditorToolbox.commands = [
+                {
+                    title: "Copy the structure of this data into the Data Structure toolbox",
+                    iconClass: "acd-icon-dataStructure",
+                    execute: (sender: IToolboxCommand) => {
+                        this.dataStructure = FieldDefinition.create(JSON.parse(this.getCurrentSampleDataEditorPayload()));
+                    }
+                }
+            ];
+
+            this._jsonEditorsPanel.addToolbox(this._sampleDataEditorToolbox);
+        }
+
+        this._jsonEditorsPanel.attachTo(document.getElementById("jsonEditorPanel"));
+
+        // Property sheet panel
+        let propertySheetHost = document.createElement("div");
+        propertySheetHost.className = "acd-propertySheet-host";
+
+        this._propertySheetToolbox = new Toolbox("propertySheet", "Element Properties");
+        this._propertySheetToolbox.content = propertySheetHost;
+
+        let propertySheetPanel = new SidePanel(
+            "propertySheetPanel",
+            SidePanelAlignment.Right,
+            document.getElementById("rightCollapsedPaneTabHost"));
+        propertySheetPanel.addToolbox(this._propertySheetToolbox);
+        propertySheetPanel.onResized = (sender: SidePanel) => {
             this.scheduleLayoutUpdate();
-
-            this._propertySheetPane.saveState();
         }
 
-        // Property sheet pane
-        this._propertySheetPane = new SidePane(
-            document.getElementById("propertySheetPane"),
-            document.getElementById("rightCollapsedPaneTabHost"),
-            "propertySheet",
-            "Element properties",
-            "selector-propertySheet");
-        this._propertySheetPane.onToggled = (sender: SidePane) => {
-            this.updateFullLayout();
-        };
-        this._propertySheetPane.content = document.getElementById("propertySheetHost");
+        propertySheetPanel.attachTo(document.getElementById("propertySheetPanel"));
 
-        // Splitter for tree view pane
-        let treeViewSplitter = new Splitter(
-            document.getElementById("treeViewSplitter"),
-            document.getElementById("treeViewPane"),
-            true);
-        treeViewSplitter.minimum = 140;
-        treeViewSplitter.onResized = (splitter: Splitter, newSize: number) => {
+        // Tree view panel
+        let treeViewHost = document.createElement("div");
+        treeViewHost.className = "acd-treeView-host";
+
+        this._treeViewToolbox = new Toolbox("treeView", "Card Structure");
+        this._treeViewToolbox.content = treeViewHost;
+
+        let treeViewPanel = new SidePanel(
+            "treeViewPanel",
+            SidePanelAlignment.Right,
+            document.getElementById("rightCollapsedPaneTabHost"));
+        treeViewPanel.addToolbox(this._treeViewToolbox);
+        treeViewPanel.onResized = (sender: SidePanel) => {
             this.scheduleLayoutUpdate();
-
-            this._treeViewPane.saveState();
         }
 
-        // Tree view pane
-        this._treeViewPane = new SidePane(
-            document.getElementById("treeViewPane"),
-            document.getElementById("rightCollapsedPaneTabHost"),
-            "treeView",
-            "Visual tree view",
-            "selector-treeView");
-        this._treeViewPane.onToggled = (sender: SidePane) => {
-            this.updateFullLayout();
-        };
-        this._treeViewPane.content = document.getElementById("treeViewHost");
+        if (Shared.GlobalSettings.previewFeaturesEnabled) {
+            let dataExplorerHost = document.createElement("div");
+            dataExplorerHost.className = "acd-treeView-host";
 
-        this._designerHostElement = document.getElementById("designerHost")
+            this._dataToolbox = new Toolbox("data", "Data Structure");
+            this._dataToolbox.content = dataExplorerHost;
 
-        this.recreateDesignerSurface();
+            treeViewPanel.addToolbox(this._dataToolbox);
+        }
 
+        treeViewPanel.attachTo(document.getElementById("treeViewPanel"));
 
-        this.loadMonaco(() => { this.monacoModuleLoaded(); });
+        this._designerHostElement = document.getElementById("designerHost");
 
         window.addEventListener("pointermove", (e: PointerEvent) => { this.handlePointerMove(e); });
         window.addEventListener("resize", () => { this.scheduleLayoutUpdate(); });
         window.addEventListener("pointerup", (e: PointerEvent) => { this.handlePointerUp(e); });
+
+        this._isAttached = true;
+
+        this.recreateDesignerSurface();
 
         let card = new Adaptive.AdaptiveCard();
         card.onImageLoaded = (image: Adaptive.Image) => {
@@ -1087,7 +1007,7 @@ export class CardDesigner {
 
             let card = this._undoStack[this._undoStackIndex];
 
-            this.setJsonPayload(card);
+            this.setCardPayload(card);
 
             this.updateToolbar();
         }
@@ -1099,7 +1019,7 @@ export class CardDesigner {
 
             let card = this._undoStack[this._undoStackIndex];
 
-            this.setJsonPayload(card);
+            this.setCardPayload(card);
 
             this.updateToolbar();
         }
@@ -1113,7 +1033,7 @@ export class CardDesigner {
             ]
         }
 
-        this.setJsonPayload(card);
+        this.setCardPayload(card);
     }
 
     setCard(payload: object) {
@@ -1135,6 +1055,38 @@ export class CardDesigner {
         return this.designerSurface.card.toJSON();
     }
 
+    onCardPayloadChanged: (designer: CardDesigner) => void;
+    onActiveHostContainerChanged: (designer: CardDesigner) => void;
+
+    get currentTargetVersion(): Shared.TargetVersion {
+        if (this._versionChoicePicker) {
+            return Shared.SupportedTargetVersions[parseInt(this._versionChoicePicker.value)];
+        }
+        else {
+            return Shared.Versions.latest;
+        }
+    }
+
+    get dataStructure(): FieldDefinition {
+        return this._dataStructure;
+    }
+
+    set dataStructure(value: FieldDefinition) {
+        this._dataStructure = value;
+
+        this.buildDataExplorer();
+    }
+
+    get sampleData(): any {
+        return this._sampleData;
+    }
+
+    set sampleData(value: any) {
+        this._sampleData = value;
+
+        this.setSampleDataPayload(value);
+    }
+    
     get activeHostContainer(): HostContainer {
         return this._activeHostContainer;
     }
@@ -1181,20 +1133,24 @@ export class CardDesigner {
         return this._designerSurface;
     }
 
-    get treeViewPane(): SidePane {
-        return this._treeViewPane;
+    get treeViewToolbox(): Toolbox {
+        return this._treeViewToolbox;
     }
 
-    get propertySheetPane(): SidePane {
-        return this._propertySheetPane;
+    get propertySheetToolbox(): Toolbox {
+        return this._propertySheetToolbox;
     }
 
-    get jsonEditorPane(): SidePane {
-        return this._jsonEditorPane;
+    get jsonEditorToolbox(): Toolbox {
+        return this._cardEditorToolbox;
     }
 
-    get toolPalettePane(): SidePane {
-        return this._toolPalettePane;
+    get toolPaletteToolbox(): Toolbox {
+        return this._toolPaletteToolbox;
+	}
+	
+    get dataToolbox(): Toolbox {
+        return this._dataToolbox;
 	}
 
 	get assetPath(): string {
@@ -1203,16 +1159,29 @@ export class CardDesigner {
 
 	set assetPath(value: string) {
 		this._assetPath = value;
-	}
+    }
+    
+    get customPaletteItems(): CustomPaletteItem[] {
+        return this._customPeletteItems;
+    }
+
+    set customPaletteItems(value: CustomPaletteItem[]) {
+        this._customPeletteItems = value;
+
+        this.buildPalette();
+    }
 }
 
 export module CardDesigner {
     export class ToolbarCommands {
         static FullScreen = "__fullScreenButton";
+        static OpenPayload = "__openPayload";
         static HostAppPicker = "__hostAppPicker";
+        static VersionPicker = "__versionPicker";
         static Undo = "__undoButton";
         static Redo = "__redoButton";
         static NewCard = "__newCardButton";
         static CopyJSON = "__copyJsonButton";
+        static TogglePreview = "__togglePreviewButton";
     }
 }
