@@ -1,11 +1,10 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import * as Adaptive from "adaptivecards";
-import * as Designer from "../../adaptivecards-designer";
+import { HostContainer } from "../host-container";
+import * as hostConfig from "../../hostConfigs/cortana-skills.json";
 
-var cortanaSkillsConfiguration = require("../../../../../../samples/HostConfig/cortana-skills.json");
-
-export class CortanaContainer extends Designer.HostContainer {
+export class CortanaContainer extends HostContainer {
     public renderTo(hostElement: HTMLElement) {
         this.cardHost.classList.add("cortana-card");
 
@@ -62,6 +61,6 @@ export class CortanaContainer extends Designer.HostContainer {
     }
 
     public getHostConfig(): Adaptive.HostConfig {
-        return new Adaptive.HostConfig(cortanaSkillsConfiguration);
+        return new Adaptive.HostConfig(hostConfig);
     }
 }
