@@ -1,11 +1,10 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import * as Adaptive from "adaptivecards";
-import * as Designer from "../../adaptivecards-designer";
+import { HostContainer } from "../host-container";
+import * as hostConfig from "../../hostConfigs/windows-notification.json";
 
-var windowsNotificationConfiguration = require("../../../../../../samples/HostConfig/windows-notification.json");
-
-export class ToastContainer extends Designer.HostContainer {
+export class ToastContainer extends HostContainer {
     public renderTo(hostElement: HTMLElement) {
         this.cardHost.classList.add("toast-card");
 
@@ -17,6 +16,6 @@ export class ToastContainer extends Designer.HostContainer {
     }
 
     public getHostConfig(): Adaptive.HostConfig {
-        return new Adaptive.HostConfig(windowsNotificationConfiguration);
+        return new Adaptive.HostConfig(hostConfig);
     }
 }
