@@ -1,11 +1,10 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import * as Adaptive from "adaptivecards";
-import * as Designer from "../../adaptivecards-designer";
+import { HostContainer } from "../host-container";
+import * as hostConfig from "../../hostConfigs/skype.json";
 
-var skypeConfiguration = require("../../../../../../samples/HostConfig/skype.json");
-
-export class SkypeContainer extends Designer.HostContainer {
+export class SkypeContainer extends HostContainer {
     public renderTo(hostElement: HTMLElement) {
         this.cardHost.classList.add("skype-card");
 
@@ -32,6 +31,6 @@ export class SkypeContainer extends Designer.HostContainer {
     }
 
     public getHostConfig(): Adaptive.HostConfig {
-        return new Adaptive.HostConfig(skypeConfiguration);
+        return new Adaptive.HostConfig(hostConfig);
     }
 }
