@@ -1,12 +1,11 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import * as Adaptive from "adaptivecards";
-import * as Designer from "../../adaptivecards-designer";
+import { HostContainer } from "../host-container";
+import * as hostConfigLight from "../../hostConfigs/microsoft-teams-light.json";
+import * as hostConfigDark from "../../hostConfigs/microsoft-teams-dark.json";
 
-var teamsDarkConfiguration = require("../../../../../../samples/HostConfig/microsoft-teams-dark.json");
-var teamsLightConfiguration = require("../../../../../../samples/HostConfig/microsoft-teams-light.json");
-
-export abstract class BaseTeamsContainer extends Designer.HostContainer {
+export abstract class BaseTeamsContainer extends HostContainer {
     public renderTo(hostElement: HTMLElement) {
         var outerFrame = document.createElement("div");
         outerFrame.className = "teams-frame";
@@ -46,7 +45,7 @@ export abstract class BaseTeamsContainer extends Designer.HostContainer {
 
 export class LightTeamsContainer extends BaseTeamsContainer {
     public getHostConfig(): Adaptive.HostConfig {
-        return new Adaptive.HostConfig(teamsLightConfiguration);
+        return new Adaptive.HostConfig(hostConfigLight);
     }
 }
 
@@ -56,6 +55,6 @@ export class DarkTeamsContainer extends BaseTeamsContainer {
     }
 
     public getHostConfig(): Adaptive.HostConfig {
-        return new Adaptive.HostConfig(teamsDarkConfiguration);
+        return new Adaptive.HostConfig(hostConfigDark);
     }
 }
