@@ -1,11 +1,10 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 import * as Adaptive from "adaptivecards";
-import * as Designer from "../../adaptivecards-designer";
+import { HostContainer } from "../host-container";
+import * as hostConfig from "../../hostConfigs/webchat.json";
 
-var webChatConfiguration = require("../../../../../../samples/HostConfig/webchat.json");
-
-export class WebChatContainer extends Designer.HostContainer {
+export class WebChatContainer extends HostContainer {
     public renderTo(hostElement: HTMLElement) {
         this.cardHost.classList.add("webChatOuterContainer");
 
@@ -17,6 +16,6 @@ export class WebChatContainer extends Designer.HostContainer {
     }
 
     public getHostConfig(): Adaptive.HostConfig {
-        return new Adaptive.HostConfig(webChatConfiguration);
+        return new Adaptive.HostConfig(hostConfig);
     }
 }
