@@ -4,6 +4,7 @@ import * as DesignerSurface from "./card-designer-surface";
 import * as DesignerPeers from "./designer-peers";
 import * as Adaptive from "adaptivecards";
 import { DraggableElement } from "./draggable-element";
+import { FieldDefinition } from "./data";
 
 export abstract class BasePaletteItem extends DraggableElement {
     protected abstract getText(): string;
@@ -15,7 +16,7 @@ export abstract class BasePaletteItem extends DraggableElement {
         element.style.display = "flex";
 
         let iconElement = document.createElement("div");
-        iconElement.classList.add("acd-icon", this.getIconClass());
+        iconElement.classList.add("acd-icon", "acd-toolPalette-icon", this.getIconClass());
         iconElement.style.flex = "0 0 auto";
 
         let labelElement = document.createElement("div");
@@ -29,7 +30,7 @@ export abstract class BasePaletteItem extends DraggableElement {
         return element;
     }
 
-    cloneElement(): HTMLElement {
+    renderDragVisual(): HTMLElement {
         return this.internalRender();
     }
 
@@ -63,26 +64,69 @@ export class ElementPaletteItem extends BasePaletteItem {
     }
 }
 
-/* DO NOT REMOVE - future feature
-class SnippetPaletteItem extends BasePaletteItem {
+export class DataPaletteItem extends BasePaletteItem {
     protected getText(): string {
-        return this.name;
+        return this.field.name;
     }
 
     protected getIconClass(): string {
         return null;
     }
 
+    constructor(readonly field: FieldDefinition) {
+        super();
+    }
+
+    createPeer(designer: DesignerSurface.CardDesignerSurface): DesignerPeers.CardElementPeer {
+        let element: Adaptive.CardElement;
+
+        if (this.field.isCollection) {
+            element = new Adaptive.Container();
+            element.setCustomProperty("$data", "{" + this.field.getPath() + "}");
+        }
+        else {
+            let textBlock = new Adaptive.TextBlock();
+            textBlock.text = "{" + this.field.getPath() + "}";
+
+            element = textBlock;
+        }
+
+        let peer = DesignerSurface.CardDesignerSurface.cardElementPeerRegistry.createPeerInstance(designer, null, element);
+        peer.initializeCardElement();
+
+        return peer;
+    }
+}
+
+export abstract class CustomPaletteItem extends BasePaletteItem {
+    readonly category: string;
+
+    protected getIconClass(): string {
+        return "acd-icon-customPaletteItem";
+    }
+
+    constructor(category: string) {
+        super();
+
+        this.category = category;
+    }
+}
+
+export class SnippetPaletteItem extends CustomPaletteItem {
+    protected getText(): string {
+        return this.name;
+    }
+
     readonly name: string;
     snippet: object;
 
-    constructor(name: string) {
-        super();
+    constructor(category: string, name: string) {
+        super(category);
 
         this.name = name;
     }
 
-    createPeer(designer: Designer.CardDesigner): Designer.CardElementPeer {
+    createPeer(designer: DesignerSurface.CardDesignerSurface): DesignerPeers.CardElementPeer {
         if (this.snippet) {
             let rootElementTypeName = this.snippet["type"];
 
@@ -92,7 +136,7 @@ class SnippetPaletteItem extends BasePaletteItem {
                 if (adaptiveElement) {
                     adaptiveElement.parse(this.snippet);
 
-                    let peer = Designer.CardDesigner.cardElementPeerRegistry.createPeerInstance(designer, null, adaptiveElement);
+                    let peer = DesignerSurface.CardDesignerSurface.cardElementPeerRegistry.createPeerInstance(designer, null, adaptiveElement);
                     peer.initializeCardElement();
 
                     return peer;
@@ -101,4 +145,3 @@ class SnippetPaletteItem extends BasePaletteItem {
         }
     }
 }
-*/
